// src/Pages/Dashboard/interfaces/Dashboard.interfaces.tsx
export interface Root {
    info: Info
    results: Character[]
  }
  
  export interface Info {
    count: number
    pages: number
    next: string
    prev: any
  }
  
  export interface Character {
    id: number
    name: string
    status: string
    species: string
    type: string
    gender: string
    origin: Origin
    location: Location
    image: string
    episode: string[]
    url: string
    created: string
  }
  
  export interface Origin {
    name: string
    url: string
  }
   
  export interface Location {
    id: number
    name: string
    type: string
    dimension: string
    residents: string[]
    url: string
    created: string
  }
  